refactoring dao profile routing
