+++ src/pages/DaoProfile/tabs/index.ts
@@ -3,3 +3,4 @@ export { default as DaoProfileTabBalance } from "./DaoProfileTabBalance"
 export { default as DaoProfileTabValidators } from "./DaoProfileTabValidators"
 export { default as DaoProfileTabDelegations } from "./DaoProfileTabDelegations"
 export { default as TabFallback } from "./TabFallback"
+export { default as Routing } from "./Routing"
